Use dedicated AffinityManager instead of general-purpose Affinity for Platform Cache needs
