+++ modules/platforms/dotnet/Apache.Ignite.Core/Impl/Ignite.cs
@@ -35,6 +35,7 @@ namespace Apache.Ignite.Core.Impl
     using Apache.Ignite.Core.DataStructures;
     using Apache.Ignite.Core.Events;
     using Apache.Ignite.Core.Impl.Binary;
+    using Apache.Ignite.Core.Impl.Binary.IO;
     using Apache.Ignite.Core.Impl.Cache;
     using Apache.Ignite.Core.Impl.Cache.Platform;
     using Apache.Ignite.Core.Impl.Cluster;
@@ -101,7 +102,8 @@ namespace Apache.Ignite.Core.Impl
             SetBaselineAutoAdjustTimeout = 35,
             GetCacheConfig = 36,
             GetThreadLocal = 37,
-            GetOrCreateLock = 38
+            GetOrCreateLock = 38,
+            GetAffinityManager = 39,
         }
 
         /** */
@@ -139,7 +141,7 @@ namespace Apache.Ignite.Core.Impl
             new ConcurrentDictionary<Guid, ClusterNodeImpl>();
 
         /** Client reconnect task completion source. */
-        private volatile TaskCompletionSource<bool> _clientReconnectTaskCompletionSource = 
+        private volatile TaskCompletionSource<bool> _clientReconnectTaskCompletionSource =
             new TaskCompletionSource<bool>();
 
         /** Plugin processor. */
@@ -189,7 +191,7 @@ namespace Apache.Ignite.Core.Impl
             SetCompactFooter();
 
             _pluginProcessor = new PluginProcessor(this);
-            
+
             _platformCacheManager = new PlatformCacheManager(this);
         }
 
@@ -470,7 +472,7 @@ namespace Apache.Ignite.Core.Impl
         public ICache<TK, TV> GetOrCreateCache<TK, TV>(CacheConfiguration configuration, NearCacheConfiguration nearConfiguration,
             PlatformCacheConfiguration platformCacheConfiguration)
         {
-            return GetOrCreateCache<TK, TV>(configuration, nearConfiguration, platformCacheConfiguration, 
+            return GetOrCreateCache<TK, TV>(configuration, nearConfiguration, platformCacheConfiguration,
                 Op.GetOrCreateCacheFromConfig);
         }
 
@@ -491,7 +493,7 @@ namespace Apache.Ignite.Core.Impl
         }
 
         /** <inheritdoc /> */
-        public ICache<TK, TV> CreateCache<TK, TV>(CacheConfiguration configuration, 
+        public ICache<TK, TV> CreateCache<TK, TV>(CacheConfiguration configuration,
             NearCacheConfiguration nearConfiguration)
         {
             return CreateCache<TK, TV>(configuration, nearConfiguration, null);
@@ -501,14 +503,14 @@ namespace Apache.Ignite.Core.Impl
         public ICache<TK, TV> CreateCache<TK, TV>(CacheConfiguration configuration, NearCacheConfiguration nearConfiguration,
             PlatformCacheConfiguration platformCacheConfiguration)
         {
-            return GetOrCreateCache<TK, TV>(configuration, nearConfiguration, platformCacheConfiguration, 
+            return GetOrCreateCache<TK, TV>(configuration, nearConfiguration, platformCacheConfiguration,
                 Op.CreateCacheFromConfig);
         }
 
         /// <summary>
         /// Gets or creates the cache.
         /// </summary>
-        private ICache<TK, TV> GetOrCreateCache<TK, TV>(CacheConfiguration configuration, 
+        private ICache<TK, TV> GetOrCreateCache<TK, TV>(CacheConfiguration configuration,
             NearCacheConfiguration nearConfiguration, PlatformCacheConfiguration platformCacheConfiguration, Op op)
         {
             IgniteArgumentCheck.NotNull(configuration, "configuration");
@@ -646,8 +648,19 @@ namespace Apache.Ignite.Core.Impl
             IgniteArgumentCheck.NotNull(cacheName, "cacheName");
 
             var aff = DoOutOpObject((int) Op.GetAffinity, w => w.WriteString(cacheName));
-            
+
             return new CacheAffinityImpl(aff, false);
+        }
+
+        /** <inheritdoc /> */
+        public CacheAffinityManager GetAffinityManager(string cacheName)
+        {
+            IgniteArgumentCheck.NotNull(cacheName, "cacheName");
+
+            var mgr = DoOutOpObject((int) Op.GetAffinityManager,
+                (IBinaryStream s) => s.WriteInt(BinaryUtils.GetCacheId(cacheName)));
+
+            return new CacheAffinityManager(mgr);
         }
 
         /** <inheritdoc /> */
@@ -918,7 +931,7 @@ namespace Apache.Ignite.Core.Impl
         public void EnableWal(string cacheName)
         {
             IgniteArgumentCheck.NotNull(cacheName, "cacheName");
-            
+
             DoOutOp((int) Op.EnableWal, w => w.WriteString(cacheName));
         }
 
@@ -933,7 +946,7 @@ namespace Apache.Ignite.Core.Impl
         /** <inheritdoc /> */
         public void SetTxTimeoutOnPartitionMapExchange(TimeSpan timeout)
         {
-            DoOutOp((int) Op.SetTxTimeoutOnPartitionMapExchange, 
+            DoOutOp((int) Op.SetTxTimeoutOnPartitionMapExchange,
                 (BinaryWriter w) => w.WriteLong((long) timeout.TotalMilliseconds));
         }
 
@@ -1005,7 +1018,7 @@ namespace Apache.Ignite.Core.Impl
             {
                 Name = name
             };
-            
+
             return GetOrCreateLock(configuration, true);
         }
 
@@ -1014,7 +1027,7 @@ namespace Apache.Ignite.Core.Impl
         {
             IgniteArgumentCheck.NotNull(configuration, "configuration");
             IgniteArgumentCheck.NotNullOrEmpty(configuration.Name, "configuration.Name");
-            
+
             // Create a copy to ignore modifications from outside.
             var cfg = new LockConfiguration(configuration);
 
@@ -1025,7 +1038,7 @@ namespace Apache.Ignite.Core.Impl
                 w.WriteBoolean(configuration.IsFair);
                 w.WriteBoolean(create);
             });
-            
+
             return target == null ? null : new IgniteLock(target, cfg);
         }
 
@@ -1123,13 +1136,13 @@ namespace Apache.Ignite.Core.Impl
         internal ITransactions GetTransactionsWithLabel(string label)
         {
             Debug.Assert(label != null);
-            
+
             var platformTargetInternal = DoOutOpObject((int) Op.GetTransactions, s =>
             {
                 var w = BinaryUtils.Marshaller.StartMarshal(s);
                 w.WriteString(label);
             });
-            
+
             return new TransactionsImpl(this, platformTargetInternal, GetLocalNode().Id, label);
         }
 
@@ -1163,7 +1176,7 @@ namespace Apache.Ignite.Core.Impl
 
             // Raise events.
             _clientReconnectTaskCompletionSource = new TaskCompletionSource<bool>();
-            
+
             var handler = ClientDisconnected;
             if (handler != null)
                 handler.Invoke(this, EventArgs.Empty);
@@ -1176,7 +1189,7 @@ namespace Apache.Ignite.Core.Impl
         internal void OnClientReconnected(bool clusterRestarted)
         {
             _marsh.OnClientReconnected(clusterRestarted);
-            
+
             _clientReconnectTaskCompletionSource.TrySetResult(clusterRestarted);
 
             var handler = ClientReconnected;
